% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stubentiger.R
\name{fit_example_to_sim_data}
\alias{fit_example_to_sim_data}
\title{Title}
\usage{
fit_example_to_sim_data(
  sim_parameters = list(v = 0.081, sigma_f = log(0.035), prop_f = 0.3, m = log(0.013)),
  steps_mcmc1 = 10,
  steps_mcmc2 = 50,
  time_steps_for_sim = 72
)
}
\arguments{
\item{sim_parameters}{list of parameter values for v, sigma_f, prop_f, m}

\item{steps_mcmc1}{number of mcmc steps for first run}

\item{steps_mcmc2}{number of mcmc steps for second run (should be a lot higher for a good fit ~10,000 but then needs to be run on an hpc and will take ~24 hours)}

\item{time_steps_for_sim}{number of time steps (in months) for simulating data}
}
\value{
fitted mcmc object
}
\description{
Title
}
