% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stubentiger.R
\name{fit_model_to_data}
\alias{fit_model_to_data}
\title{Title}
\usage{
fit_model_to_data(data, fixed_parameters, steps_mcmc1 = 10, steps_mcmc2 = 50)
}
\arguments{
\item{data}{data to fit model to (GPSC counts)}

\item{fixed_parameters}{input parameters from data, such as start population}

\item{steps_mcmc1}{number of mcmc steps for first run}

\item{steps_mcmc2}{number of mcmc steps for second run (should be a lot higher for a good fit ~10,000 but then needs to be run on an hpc and will take ~24 hours)}
}
\value{
fitted mcmc object
}
\description{
Title
}
