% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stubentiger.R
\name{likelihood}
\alias{likelihood}
\title{Title}
\usage{
likelihood(state, observed, pars = NULL)
}
\arguments{
\item{state}{values of model}

\item{observed}{values of data}

\item{pars}{parameters}
}
\value{
likelihood
}
\description{
Title
}
