% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stubentiger.R
\name{simulate_example_serotypes}
\alias{simulate_example_serotypes}
\title{Title}
\usage{
simulate_example_serotypes(
  sim_parameters = list(v = 0.081, sigma_f = log(0.035), prop_f = 0.3, m = log(0.013)),
  time_steps = 36
)
}
\arguments{
\item{sim_parameters}{list of parameter values for v, sigma_f, prop_f, m}

\item{time_steps}{number of time steps (in months) to run the model for}
}
\value{
simulated serotype data
}
\description{
Title
}
