% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stubentiger.R
\name{simulate_model_serotypes}
\alias{simulate_model_serotypes}
\title{Title}
\usage{
simulate_model_serotypes(model_parameters, simulation_steps)
}
\arguments{
\item{model_parameters}{list of model parameters}

\item{simulation_steps}{number of steps to run the simulation for (in model time) (if dt = 1/12, this should be 12 * years to run simulation for)}
}
\value{
simulated data in serotype frequencies over time as data frame
}
\description{
Title
}
