% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stubentiger.R
\docType{data}
\name{simulated_data_over_time_filtered}
\alias{simulated_data_over_time_filtered}
\title{Title}
\format{
A data frame with simulated_data_over_time_filtered
}
\description{
simulated_data_over_time_filtered - a dataframe used in package examples.
}
